+++ frontend/src/api/v3/admin.js
@@ -102,6 +102,10 @@ const librarianTask = (task, library_id) => {
   return HTTP.post("/admin/librarian/task", { task, library_id });
 };
 
+const getStats = () => {
+  return HTTP.get("/admin/stats");
+};
+
 export default {
   changeUserPassword,
   createGroup,
@@ -116,6 +120,7 @@ export default {
   getGroups,
   getLibrarianStatuses,
   getLibraries,
+  getStats,
   getUsers,
   librarianTask,
   postLibrarianTask,